+++ src/main/java/com/uspray/uspray/controller/GroupPrayController.java
@@ -9,7 +9,7 @@ import com.uspray.uspray.DTO.pray.PrayListResponseDto;
 import com.uspray.uspray.DTO.pray.request.PrayToGroupPrayDto;
 import com.uspray.uspray.DTO.pray.response.PrayResponseDto;
 import com.uspray.uspray.exception.SuccessStatus;
-import com.uspray.uspray.service.GroupPrayFacade;
+import com.uspray.uspray.service.facade.GroupPrayFacade;
 import com.uspray.uspray.service.GroupPrayService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
