+++ src/main/java/com/uspray/uspray/service/facade/PrayFacade.java
@@ -1,4 +1,4 @@
-package com.uspray.uspray.service;
+package com.uspray.uspray.service.facade;
 
 import com.uspray.uspray.DTO.pray.PrayListResponseDto;
 import com.uspray.uspray.DTO.pray.request.PrayRequestDto;
@@ -20,6 +20,7 @@ import com.uspray.uspray.infrastructure.MemberRepository;
 import com.uspray.uspray.infrastructure.NotificationLogRepository;
 import com.uspray.uspray.infrastructure.PrayRepository;
 import com.uspray.uspray.infrastructure.ScrapAndHeartRepository;
+import com.uspray.uspray.service.FCMNotificationService;
 import java.time.LocalDate;
 import java.util.List;
 import javax.transaction.Transactional;
